# Vectorized Discretization