### Summary / Description

The purpose of this pr is to vectorize our discretization implementation. Goal was to maximize performance gains with minimal implementation change, while also making it easy to add other reduction ops in the future (e.g. _mean_, _last_, _random_).

**Related Issues:** #90 

#### Latency

| Dataset | Main | perf/discretize | Speedup |
|:---:|:---:|:---:|:---:|
| tgbl-wiki | 0.84s | 0.02s | 29.41x |
| tgbn-genre | 95.29s | 1.06s | 89.90x |
| tgbl-coin | 123.39s | 1.28s | 96.27x |

#### Correctness
Discretization unit tests pass. Here are test results on all methods which use `.discretize()`

| Task / Model | Main | perf/discretize |
|---|---|---|
| Graph Prop Persistant Forecast | 0.6378 AUROC | 0.6378 AUROC |
| Graph Prop TGCN | 0.9087 AUROC | 0.9089 AUROC |
| Link Prop GC-LSTM (10 epochs) | 0.2283 MRR | 0.2283 MRR |
| Link Prop GCN (10 epochs) | 0.4068 MRR | 0.4041 MRR |